refactor #57: skeleton 파일 경로 수정 및 dark mode color 변경
