--- src/app/loading.js
@@ -1,13 +1,0 @@
-import TitleSkeleton from '@/components/common/TitleSkeleton';
-import PostSkeleton from '@/components/common/PostSkeleton';
-
-const Loading = () => {
-  return (
-    <>
-      <TitleSkeleton />
-      <PostSkeleton />
-    </>
-  );
-};
-
-export default Loading;
