--- src/app/snippets/loading.js
@@ -1,13 +1,0 @@
-import SnippetSkeleton from '@/components/common/SnippetSkeleton';
-import TitleSkeleton from '@/components/common/TitleSkeleton';
-
-const Loading = () => {
-  return (
-    <>
-      <TitleSkeleton />
-      <SnippetSkeleton />
-    </>
-  );
-};
-
-export default Loading;
