--- src/components/common/TitleSkeleton.tsx
@@ -1,10 +1,0 @@
-const TitleSkeleton = () => {
-  return (
-    <div className="flex flex-col space-y-2 my-10">
-      <div className="w-1/5 h-8 rounded bg-basic-150"></div>
-      <div className="w-3/5 h-10 rounded bg-basic-300"></div>
-    </div>
-  );
-};
-
-export default TitleSkeleton;
